📄 requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
cloudinary==1.39.0
Werkzeug==3.0.1
